requirements.txt
langchain
langchain-core
langchain-community
openai
pypdf
pymupdf
ipykernel
dotenv
langchain_text_splitters
pandas
numpy
sentence_transformers
faiss-cpu
pickle
